Restructure overall.
