+++ Sources/HasTermLevelQuery.swift
@@ -1,22 +1,20 @@
 //
-//  NestedQuery.swift
+//  HasTermLevelQuery.swift
 //  ElastiQ
 //
-//  Created by suguru-kishimoto on 2017/08/23.
+//  Created by suguru-kishimoto on 2017/09/04.
 //  Copyright © 2017年 Suguru Kishimoto. All rights reserved.
 //
 
 import Foundation
 
-public protocol HasParameter: class {
-    func add<T: QueryParameter>(_ parameter: T)
-    func add<T: QueryParameter>(_ parameter: T, configurationBlock: ParameterConfigurationBlock<T>)
+public protocol HasTermLevelQuery {
 }
 
-extension HasParameter {
+extension HasTermLevelQuery where Self: HasParameter {
     @discardableResult
     public func term(_ key: String, _ value: QueryValue) -> Self {
-        add(ElastiQ.Parameter.Term(key: key, value: value))
+        add(ElastiQ.TermLevelQueries.Term(key: key, value: value))
         return self
     }
 
@@ -32,7 +30,7 @@ extension HasParameter {
 
     @discardableResult
     public func terms(_ key: String, _ values: [QueryValue]) -> Self {
-        add(ElastiQ.Parameter.Terms(key: key, values: values))
+        add(ElastiQ.TermLevelQueries.Terms(key: key, values: values))
         return self
     }
 
@@ -47,33 +45,33 @@ extension HasParameter {
     }
 
     @discardableResult
-    public func range(_ key: String, _ condition: ElastiQ.Parameter.Range.Condition) -> Self {
+    public func range(_ key: String, _ condition: ElastiQ.TermLevelQueries.Range.Condition) -> Self {
         return range(key, [condition])
     }
 
     @discardableResult
-    public func range<K, V: QueryNumberValue>(_ keypath: KeyPath<K, V>, _ condition: ElastiQ.Parameter.Range.Condition) -> Self {
+    public func range<K, V: QueryNumberValue>(_ keypath: KeyPath<K, V>, _ condition: ElastiQ.TermLevelQueries.Range.Condition) -> Self {
         return range(keypath._kvcKeyPathString!, condition)
     }
 
     @discardableResult
-    public func range<K, V: QueryNumberValue>(_ keypath: KeyPath<K, V?>, _ condition: ElastiQ.Parameter.Range.Condition) -> Self {
+    public func range<K, V: QueryNumberValue>(_ keypath: KeyPath<K, V?>, _ condition: ElastiQ.TermLevelQueries.Range.Condition) -> Self {
         return range(keypath._kvcKeyPathString!, condition)
     }
 
     @discardableResult
-    public func range(_ key: String, _ conditions: [ElastiQ.Parameter.Range.Condition]) -> Self {
-        add(ElastiQ.Parameter.Range(key: key, conditions: conditions))
+    public func range(_ key: String, _ conditions: [ElastiQ.TermLevelQueries.Range.Condition]) -> Self {
+        add(ElastiQ.TermLevelQueries.Range(key: key, conditions: conditions))
         return self
     }
 
     @discardableResult
-    public func range<K, V: QueryNumberValue>(_ keypath: KeyPath<K, V>, _ conditions: [ElastiQ.Parameter.Range.Condition]) -> Self {
+    public func range<K, V: QueryNumberValue>(_ keypath: KeyPath<K, V>, _ conditions: [ElastiQ.TermLevelQueries.Range.Condition]) -> Self {
         return range(keypath._kvcKeyPathString!, conditions)
     }
 
     @discardableResult
-    public func range<K, V: QueryNumberValue>(_ keypath: KeyPath<K, V?>, _ conditions: [ElastiQ.Parameter.Range.Condition]) -> Self {
+    public func range<K, V: QueryNumberValue>(_ keypath: KeyPath<K, V?>, _ conditions: [ElastiQ.TermLevelQueries.Range.Condition]) -> Self {
         return range(keypath._kvcKeyPathString!, conditions)
     }
 }
