+++ Sources/Query.swift
@@ -9,7 +9,7 @@
 import Foundation
 
 extension ElastiQ {
-    public final class Query: HasSingleParameter, QueryParameter {
+    public final class Query: HasSingleParameter, QueryParameter, HasTermLevelQuery {
         public let parameterName = "query"
 
         public var parameter: QueryParameter?
